requirements.txt
Django==4.2
djangorestframework==3.16.0
mysqlclient==2.2.7